Split query helpers into separate module
